KTAC/requirements.txt
torch>=2.0.0
transformers>=4.35.0
einops>=0.7.0
timm>=0.9.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
h5py>=3.9.0
pyyaml>=6.0
tabulate>=0.9.0
